Move BaseAtomsDataModule
