+++ src/schnetpack/datasets/__init__.py
@@ -1,1 +1,2 @@
+from .base import AtomsDataModule
 from .qm9 import QM9
